apps/discord-bot/requirements.txt
# Discord Bot Core
discord.py>=2.3.0
asyncpg>=0.29.0
python-dotenv>=1.0.0

# AI
openai>=1.12.0
PyPDF2>=3.0.0

# Chart
matplotlib>=3.8.0

# Redis
redis[hiredis]>=5.0.0

# HTTP
aiohttp>=3.9.0
